feat: Align module with Snowflake provider changes
Due to breaking changes in Snowflake provider and additional code optimizations, **breaking changes** were introduced in this module.

List of code and variable (API) changes:
- Switched to `snowflake_database_role` module to leverage new `database_roles` mechanism
- `default_roles` and `custom_roles` are now combined and controlled by single logic in snowflake_database_role module
- snowflake_database resource was updated to use newly introduced changes in Snowflake provider
- snowflake_schema resource was updated to use newly introduced changes in Snowflake provider
- variable `add_grants_to_existing_objects` was removed as it is no longer needed
- minimum Snowflake provider version is `0.90.0`

For more information, refer to [variables.tf](variables.tf), list of inputs below and Snowflake provider documentation

When upgrading from `v1.x`, expect most of the resources to be recreated - if recreation is impossible, then it is possible to import some existing resources.